clean up linked list and add new block storage layout in comments
