+++ pg_search/src/postgres/storage/linked_list.rs
@@ -15,7 +15,9 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-use super::block::{bm25_max_free_space, BM25PageSpecialData, LinkedListData, MVCCEntry, PgItem};
+use super::block::{
+    bm25_max_free_space, BM25PageSpecialData, LinkedList, LinkedListData, MVCCEntry, PgItem,
+};
 use super::utils::{BM25BufferCache, BM25Page};
 use anyhow::{bail, Result};
 use pgrx::pg_sys;
@@ -31,6 +33,26 @@ pub struct LinkedItemList<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCE
     pub lock_buffer: pg_sys::Buffer,
     lock: Option<u32>,
     _marker: std::marker::PhantomData<T>,
+}
+
+impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> LinkedList for LinkedItemList<T> {
+    fn get_lock_buffer(&self) -> pg_sys::Buffer {
+        self.lock_buffer
+    }
+
+    fn get_lock(&self) -> Option<u32> {
+        self.lock
+    }
+}
+
+impl LinkedList for LinkedBytesList {
+    fn get_lock_buffer(&self) -> pg_sys::Buffer {
+        self.lock_buffer
+    }
+
+    fn get_lock(&self) -> Option<u32> {
+        self.lock
+    }
 }
 
 impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> LinkedItemList<T> {
@@ -76,6 +98,11 @@ impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> LinkedItemList<
     }
 
     pub unsafe fn garbage_collect(&mut self) -> Result<()> {
+        assert!(
+            self.lock == Some(pg_sys::BUFFER_LOCK_EXCLUSIVE),
+            "an exclusive lock is required to garbage collect"
+        );
+
         // Collect all entries that are not definitely deleted
         let mut entries_to_keep = vec![];
         let snapshot = pg_sys::GetActiveSnapshot();
@@ -93,7 +120,7 @@ impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> LinkedItemList<
 
         // Mark all buffer as deleted besides the lock buffer
         let cache = BM25BufferCache::open(self.relation_oid);
-        let mut blockno = get_start_blockno(self.lock_buffer);
+        let mut blockno = self.get_start_blockno();
         while blockno != pg_sys::InvalidBlockNumber {
             let buffer = cache.get_buffer(blockno, Some(pg_sys::BUFFER_LOCK_EXCLUSIVE));
             let page = pg_sys::BufferGetPage(buffer);
@@ -245,18 +272,6 @@ impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> LinkedItemList<
 
         Ok(items)
     }
-
-    unsafe fn get_start_blockno(&self) -> pg_sys::BlockNumber {
-        get_start_blockno(self.lock_buffer)
-    }
-
-    unsafe fn get_last_blockno(&self) -> pg_sys::BlockNumber {
-        get_last_blockno(self.lock_buffer)
-    }
-
-    unsafe fn set_last_blockno(&self, blockno: pg_sys::BlockNumber) {
-        set_last_blockno(self.lock_buffer, self.lock, blockno);
-    }
 }
 
 impl<T: From<PgItem> + Into<PgItem> + Debug + Clone + MVCCEntry> Drop for LinkedItemList<T> {
@@ -394,7 +409,7 @@ impl LinkedBytesList {
 
     pub unsafe fn mark_deleted(&self) {
         let cache = BM25BufferCache::open(self.relation_oid);
-        let mut blockno = get_start_blockno(self.lock_buffer);
+        let mut blockno = self.get_start_blockno();
         while blockno != pg_sys::InvalidBlockNumber {
             let buffer = cache.get_buffer(blockno, Some(pg_sys::BUFFER_LOCK_EXCLUSIVE));
             let page = pg_sys::BufferGetPage(buffer);
@@ -425,19 +440,19 @@ impl LinkedBytesList {
 
     pub unsafe fn get_all_blocks(&self) -> Vec<pg_sys::BlockNumber> {
         let cache = BM25BufferCache::open(self.relation_oid);
-        get_all_blocks(self.lock_buffer, &cache)
-    }
+        let mut blockno = self.get_start_blockno();
+        let mut blocks = vec![];
 
-    unsafe fn get_start_blockno(&self) -> pg_sys::BlockNumber {
-        get_start_blockno(self.lock_buffer)
-    }
+        while blockno != pg_sys::InvalidBlockNumber {
+            blocks.push(blockno);
+            let buffer = cache.get_buffer(blockno, Some(pg_sys::BUFFER_LOCK_SHARE));
+            let page = pg_sys::BufferGetPage(buffer);
+            let special = pg_sys::PageGetSpecialPointer(page) as *mut BM25PageSpecialData;
+            blockno = (*special).next_blockno;
+            pg_sys::UnlockReleaseBuffer(buffer);
+        }
 
-    unsafe fn get_last_blockno(&self) -> pg_sys::BlockNumber {
-        get_last_blockno(self.lock_buffer)
-    }
-
-    unsafe fn set_last_blockno(&self, blockno: pg_sys::BlockNumber) {
-        set_last_blockno(self.lock_buffer, self.lock, blockno);
+        blocks
     }
 }
 
@@ -453,71 +468,6 @@ impl Drop for LinkedBytesList {
             }
         };
     }
-}
-
-#[inline]
-unsafe fn get_list_metadata(lock_buffer: pg_sys::Buffer) -> LinkedListData {
-    let page = pg_sys::BufferGetPage(lock_buffer);
-    let metadata = pg_sys::PageGetContents(page) as *mut LinkedListData;
-    LinkedListData {
-        start_blockno: (*metadata).start_blockno,
-        last_blockno: (*metadata).last_blockno,
-    }
-}
-
-#[inline]
-fn get_start_blockno(lock_buffer: pg_sys::Buffer) -> pg_sys::BlockNumber {
-    let metadata = unsafe { get_list_metadata(lock_buffer) };
-    let start_blockno = metadata.start_blockno;
-    assert!(start_blockno != 0);
-    assert!(start_blockno != pg_sys::InvalidBlockNumber);
-    start_blockno
-}
-
-#[inline]
-fn get_last_blockno(lock_buffer: pg_sys::Buffer) -> pg_sys::BlockNumber {
-    let metadata = unsafe { get_list_metadata(lock_buffer) };
-    let last_blockno = metadata.last_blockno;
-    assert!(last_blockno != 0);
-    assert!(last_blockno != pg_sys::InvalidBlockNumber);
-    last_blockno
-}
-
-#[inline]
-unsafe fn set_last_blockno(
-    lock_buffer: pg_sys::Buffer,
-    lock: Option<u32>,
-    blockno: pg_sys::BlockNumber,
-) {
-    assert!(
-        lock == Some(pg_sys::BUFFER_LOCK_EXCLUSIVE),
-        "an exclusive lock is required to write to linked list"
-    );
-
-    let page = pg_sys::BufferGetPage(lock_buffer);
-    let metadata = pg_sys::PageGetContents(page) as *mut LinkedListData;
-    (*metadata).last_blockno = blockno;
-    pg_sys::MarkBufferDirty(lock_buffer);
-}
-
-#[inline]
-unsafe fn get_all_blocks(
-    lock_buffer: pg_sys::Buffer,
-    cache: &BM25BufferCache,
-) -> Vec<pg_sys::BlockNumber> {
-    let mut blockno = get_start_blockno(lock_buffer);
-    let mut blocks = vec![];
-
-    while blockno != pg_sys::InvalidBlockNumber {
-        blocks.push(blockno);
-        let buffer = cache.get_buffer(blockno, Some(pg_sys::BUFFER_LOCK_SHARE));
-        let page = pg_sys::BufferGetPage(buffer);
-        let special = pg_sys::PageGetSpecialPointer(page) as *mut BM25PageSpecialData;
-        blockno = (*special).next_blockno;
-        pg_sys::UnlockReleaseBuffer(buffer);
-    }
-
-    blocks
 }
 
 #[cfg(any(test, feature = "pg_test"))]
